Restructure source code into src/model_config_tests and add pyproject.toml
