+++ .gitignore
@@ -1,0 +1,10 @@
+*.pyc
+/__pycache__
+/build
+/dist
+*.egg-info
+.coverage
+.ipynb_checkpoints
+.vscode
+.idea
+/tests/tmp/
