--- models/__init__.py
@@ -1,5 +1,0 @@
-from models.accessom2 import AccessOm2
-
-index = {
-    'access-om2': AccessOm2
-}
