--- requirements.txt
@@ -1,6 +1,0 @@
-pytest==8.0.1
-jsonschema==4.21.1
-requests
-PyYAML
-f90nml>=0.16
-ruamel.yaml>=0.18.5
