+++ src/model_config_tests/models/__init__.py
@@ -1,0 +1,5 @@
+from model_config_tests.models.accessom2 import AccessOm2
+
+index = {
+    'access-om2': AccessOm2
+}
